Use annotations from desktop app
This comparison has been run in the original desktop app
and should score 15.12, but is actually scoring 5302.94

The fingerprint files are included here as fgp1.fgp and fgp2.fgp.
The reference and spot points were extracted from these using
https://www.npmjs.com/package/i3s-fgp-reader

So, either there's a problem with the way we're passing data to
the algorithm code, or the above coordinate extraction is
incorrect.